requirements.txt
Flask==3.1.2
numpy==2.3.5
tensorflow-cpu==2.20.0
h5py==3.15.1
joblib==1.5.2
Pillow==12.0.0
matplotlib==3.10.7
tifffile==2025.10.16
requests==2.32.5
huggingface_hub==0.17.1
gunicorn==21.2.0
